app/backend/apps/auth/serializers.py
from rest_framework import serializers
from django.contrib.auth.models import User


class UserSerializer(serializers.ModelSerializer):
    """ユーザーシリアライザー"""
    
    class Meta:
        model = User
        fields = ('id', 'username', 'email')
        read_only_fields = ('id',)

